+++ dodola/tests/test_repository.py
@@ -1,3 +1,5 @@
+import json
+import fsspec
 from xarray import Dataset, open_zarr
 import dodola.repository
 
@@ -14,3 +16,15 @@ def test_memory_repository_write():
     url = "memory://test_memory_repository_write.zarr"
     dodola.repository.write(url, Dataset({"bar": "SPAM"}))
     assert open_zarr(url) == Dataset({"bar": "SPAM"})
+
+
+def test_read_attrs():
+    """Basic test that read_attrs reads and deserializes JSON"""
+    url = "memory://test_read_attrs.json"
+    payload = {"foo": "bar"}
+    with fsspec.open(url, mode="w") as fl:
+        json.dump(payload, fl)
+
+    victim = dodola.repository.read_attrs(url)
+
+    assert victim == payload